+++ lib/src/autocomplete/autocomplete.dart
@@ -102,13 +102,7 @@ class Autocomplete {
       strictbounds,
     );
 
-    var uri = Uri.https(
-      proxyUrl != null && proxyUrl != '' ? proxyUrl! : _authority,
-      proxyUrl != null && proxyUrl != ''
-          ? 'https://$_authority/$_unencodedPath'
-          : _unencodedPath,
-      queryParameters,
-    );
+    var uri = NetworkUtility.createUri(proxyUrl, _authority, _unencodedPath, queryParameters);
     var response = await NetworkUtility.fetchUrl(uri, headers: headers);
     if (response != null) {
       return AutocompleteResponse.parseAutocompleteResult(response);
@@ -206,9 +200,7 @@ class Autocomplete {
 
     var uri = Uri.https(
       proxyUrl != null && proxyUrl != '' ? proxyUrl! : _authority,
-      proxyUrl != null && proxyUrl != ''
-          ? 'https://$_authority/$_unencodedPath'
-          : _unencodedPath,
+      proxyUrl != null && proxyUrl != '' ? 'https://$_authority/$_unencodedPath' : _unencodedPath,
       queryParameters,
     );
     return await NetworkUtility.fetchUrl(uri, headers: headers);