+++ lib/src/search/search.dart
@@ -10,8 +10,7 @@ import 'package:google_place/src/utils/network_utility.dart';
 
 class Search {
   static final _authority = 'maps.googleapis.com';
-  static final _unencodedPathFindPlace =
-      'maps/api/place/findplacefromtext/json';
+  static final _unencodedPathFindPlace = 'maps/api/place/findplacefromtext/json';
   static final _unencodedPathNearBySearch = 'maps/api/place/nearbysearch/json';
   static final _unencodedPathTextSearch = 'maps/api/place/textsearch/json';
   final String apiKEY;
@@ -57,13 +56,7 @@ class Search {
       locationbias,
     );
 
-    var uri = Uri.https(
-      proxyUrl != null && proxyUrl != '' ? proxyUrl! : _authority,
-      proxyUrl != null && proxyUrl != ''
-          ? 'https://$_authority/$_unencodedPathFindPlace'
-          : _unencodedPathFindPlace,
-      queryParameters,
-    );
+    var uri = NetworkUtility.createUri(proxyUrl, _authority, _unencodedPathFindPlace, queryParameters);
     var response = await NetworkUtility.fetchUrl(uri, headers: headers);
     if (response != null) {
       return FindPlaceResponse.parseFindPlaceResult(response);
@@ -138,8 +131,7 @@ class Search {
       pagetoken,
     );
 
-    var uri =
-        Uri.https(_authority, _unencodedPathNearBySearch, queryParameters);
+    var uri = Uri.https(_authority, _unencodedPathNearBySearch, queryParameters);
     var response = await NetworkUtility.fetchUrl(uri, headers: headers);
     if (response != null) {
       return NearBySearchResponse.parseNearBySearchResult(response);
@@ -341,8 +333,7 @@ class Search {
       pagetoken,
     );
 
-    var uri =
-        Uri.https(_authority, _unencodedPathNearBySearch, queryParameters);
+    var uri = Uri.https(_authority, _unencodedPathNearBySearch, queryParameters);
     return await NetworkUtility.fetchUrl(uri, headers: headers);
   }
 
@@ -463,8 +454,7 @@ class Search {
         value = 'ipbias';
       }
       if (locationbias.point != null) {
-        value =
-            'point:${locationbias.point!.latitude},${locationbias.point!.longitude}';
+        value = 'point:${locationbias.point!.latitude},${locationbias.point!.longitude}';
       }
       if (locationbias.circular != null) {
         value =